+++ src/main/java/com/robertx22/age_of_exile/player_skills/items/foods/FoodType.java
@@ -7,9 +7,10 @@ import net.minecraft.item.Item;
 import net.minecraft.item.Items;
 
 public enum FoodType {
+
     APPLE("Apple", "apple", StatusEffects.WATER_BREATHING, Items.APPLE, Items.COOKED_SALMON),
     BEER("Beer", "beer", ModRegistry.POTIONS.KNOCKBACK_RESISTANCE, Items.HAY_BLOCK, Items.COOKED_BEEF),
-    JAM("Jam", "jam", StatusEffects.FIRE_RESISTANCE, Items.HONEY_BOTTLE, Items.COOKED_SALMON),
+    JAM("Jam", "jam", StatusEffects.FIRE_RESISTANCE, Items.MELON, Items.COOKED_SALMON),
     PIE("Pie", "pie", StatusEffects.HASTE, Items.PUMPKIN_PIE, Items.PUMPKIN_PIE),
     COOKIE("Cookie", "cookie", StatusEffects.SPEED, Items.COOKIE, Items.COOKED_CHICKEN),
     BREW("Brew", "brew", StatusEffects.NIGHT_VISION, Items.MUSHROOM_STEW, Items.COOKED_SALMON);